fftools/ffmpeg_filter: Move some declaration to new header file
to allow filtergraph printing to access the information.
